# docs-backup-20260213-150302/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ATO Tax Optimizer API
  version: 0.1.0
  description: >
    API for the Australian Tax Optimizer platform. Integrates with Xero for
    financial data, uses Gemini AI for forensic analysis, and provides tax
    optimisation recommendations for Australian businesses.
  contact:
    name: ATO Tax Optimizer
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://ato-app.vercel.app
    description: Production (Vercel)
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/xero/callback:
    get:
      tags: [auth]
      summary: Xero OAuth callback
      description: Handles OAuth callback from Xero and stores tokens.
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to dashboard
...
  /api/analysis/queue/status:
    get:
      tags: [analysis]
      summary: Get analysis queue status
      description: Returns the current state of the analysis processing queue.
      parameters:
        - $ref: '#/components/parameters/tenantIdQuery'
      responses:
        '200':
          description: Queue status

  # ── Reports ──
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token from authentication
  parameters:
    tenantIdQuery:
      name: tenantId
      in: query
      required: true
      description: Xero tenant ID
      schema:
        type: string
...